+++ sphinx/util/typing.py
@@ -127,6 +127,8 @@ InventoryItem = tuple[
     str,  # URL
     str,  # display name
 ]
+
+# referencable role -> (reference name -> inventory item)
 Inventory = dict[str, dict[str, InventoryItem]]