+++ CHANGES.rst
@@ -60,6 +60,9 @@ Incompatible changes
 * #12597: Change the default of :confval:`show_warning_types`
   from ``False`` to ``True``.
   Patch by Chris Sewell.
+* #12083: Remove support for the old (2008--2010) Sphinx 0.5 and Sphinx 0.6
+  :confval:`intersphinx_mapping` format.
+  Patch by Bénédikt Tran and Adam Turner.
 
 Deprecated
 ----------